split up os detection for linting
